# executor: fix unreasonable CPU time in HashJoin probe phash