--- a/executor/hash_table.go
+++ b/executor/hash_table.go
@@ -241,10 +241,14 @@ func (c *hashRowContainer) GetMatchedRowsAndPtrs(probeKey uint64, probeRow chunk
 		matchedDataSize                  = int64(cap(matched))*rowSize + int64(cap(matchedPtrs))*rowPtrSize
 		lastChunkBufPointer *chunk.Chunk = nil
 		memDelta            int64        = 0
+		trackMemUsage                    = cap(innerPtrs) > signalCheckpointForJoin
 	)
 	c.chkBuf = nil
-	c.memTracker.Consume(-c.chkBufSizeForOneProbe + int64(cap(innerPtrs))*rowPtrSize)
-	defer c.memTracker.Consume(-int64(cap(innerPtrs))*rowPtrSize + memDelta)
+	c.memTracker.Consume(-c.chkBufSizeForOneProbe)
+	if trackMemUsage {
+		c.memTracker.Consume(int64(cap(innerPtrs)) * rowPtrSize)
+		defer c.memTracker.Consume(-int64(cap(innerPtrs))*rowPtrSize + memDelta)
+	}
 	c.chkBufSizeForOneProbe = 0
 
 	for i, ptr := range innerPtrs {
@@ -257,13 +261,13 @@ func (c *hashRowContainer) GetMatchedRowsAndPtrs(probeKey uint64, probeRow chunk
 		if err != nil {
 			return nil, nil, err
 		}
-		if c.chkBuf != lastChunkBufPointer && lastChunkBufPointer != nil {
+		if trackMemUsage && c.chkBuf != lastChunkBufPointer && lastChunkBufPointer != nil {
 			lastChunkSize := lastChunkBufPointer.MemoryUsage()
 			c.chkBufSizeForOneProbe += lastChunkSize
 			memDelta += lastChunkSize
 		}
 		lastChunkBufPointer = c.chkBuf
-		if i&signalCheckpointForJoin == 0 {
+		if trackMemUsage && (i&signalCheckpointForJoin == (signalCheckpointForJoin - 1)) {
 			// Trigger Consume for checking the OOM Action signal
 			memDelta += int64(cap(matched))*rowSize + int64(cap(matchedPtrs))*rowPtrSize - matchedDataSize
 			matchedDataSize = int64(cap(matched))*rowSize + int64(cap(matchedPtrs))*rowPtrSize