@dbsid 2023-02-10
elapse time for tpch q9  50g on 3tikv
before this pr: 346s
after this pr: 111s